fix up pre-commit issues, merge duplicate test files
